+++ src/frontend/src/features/rooms/livekit/components/controls/Device/VideoDeviceControl.tsx
@@ -4,7 +4,7 @@ import { Button, Popover } from '@/primitives'
 import { RiArrowUpSLine, RiVideoOffLine, RiVideoOnLine } from '@remixicon/react'
 import { Track, VideoCaptureOptions } from 'livekit-client'
 
-import { ToggleDevice } from '@/features/rooms/livekit/components/controls/ToggleDevice'
+import { ToggleDevice } from './ToggleDevice'
 import { css } from '@/styled-system/css'
 import { usePersistentUserChoices } from '../../../hooks/usePersistentUserChoices'
 import { useCannotUseDevice } from '../../../hooks/useCannotUseDevice'
